Error when column from remote table doesn't map to foreign table
Previously, the code would set col_position_mask to -1 (line 1642),
which would later cause code at line 1682 to skip over the critical
initialization of the values array (from line 1668), which would later
cause a segfault when BuildTupleFromCStrings() (line 1824) is hit.

Instead, just avoid all of this messiness and error out if the remote
table is returning columns that we cannot map to our foreign table.

It appears that the code always queries the remote table for all columns
specified by the foreign table, regardless of the query. Thus, every
column should map 1:1.